elementary_log_based/requirements.txt
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.1
pandas==2.1.4
networkx==3.2.1
neo4j==5.16.0
python-dotenv==1.0.0
matplotlib==3.8.2
plotly==5.17.0
pyvis==0.3.2
sqlalchemy>=1.4.0,<2.0.0
jupyter==1.0.0
PyYAML==6.0.1
streamlit==1.32.0
streamlit-agraph==0.0.45